Alteração da MainFrame. 'SubFrames' adicionadas.
